requirements.txt
Pillow>=10.0.0
imagehash>=4.3.1
flask>=3.0.0
requests>=2.31.0
python-dateutil>=2.8.2
